improve factory and clean `get_name` function that was in utils but only used once
