+++ idunn/places/poi.py
@@ -4,7 +4,6 @@ from urllib.parse import urlencode
 from .base import BasePlace
 from idunn import settings
 from idunn.api.constants import PoiSource
-from ..utils.verbosity import get_name
 
 OSM_CONTRIBUTION_HASHTAGS = settings["OSM_CONTRIBUTION_HASHTAGS"]
 
@@ -22,7 +21,29 @@ class POI(BasePlace):
         return self.properties.get("name", "")
 
     def get_name(self, lang):
-        return get_name(self.properties, lang)
+        """
+        Return the Place name from the properties field of the elastic response. Here 'name'
+        corresponds to the POI name in the language of the user request (i.e. 'name:{lang}' field).
+
+        If lang is None or if name:lang is not in the properties then name receives the local name
+        value.
+
+        >>> get_name({}, 'fr') is None
+        True
+
+        >>> get_name({'name':'spontini', 'name:en':'spontinien', 'name:fr':'spontinifr'}, None)
+        'spontini'
+
+        >>> get_name({'name':'spontini', 'name:en':'spontinien', 'name:fr':'spontinifr'}, 'cz')
+        'spontini'
+
+        >>> get_name({'name':'spontini', 'name:en':'spontinien', 'name:fr':'spontinifr'}, 'fr')
+        'spontinifr'
+        """
+        name = self.properties.get(f"name:{lang}")
+        if name is None:
+            name = self.properties.get("name")
+        return name
 
     def get_class_name(self):
         return self.properties.get("poi_class")
@@ -82,3 +103,29 @@ class BragiPOI(POI):
 
     def get_country_codes(self):
         return [c.upper() for c in self.get_raw_address().get("country_codes") or []]
+
+
+def get_name(properties, lang):
+    """
+    Return the Place name from the properties field of the elastic response. Here 'name'
+    corresponds to the POI name in the language of the user request (i.e. 'name:{lang}' field).
+
+    If lang is None or if name:lang is not in the properties then name receives the local name
+    value.
+
+    >>> get_name({}, 'fr') is None
+    True
+
+    >>> get_name({'name':'spontini', 'name:en':'spontinien', 'name:fr':'spontinifr'}, None)
+    'spontini'
+
+    >>> get_name({'name':'spontini', 'name:en':'spontinien', 'name:fr':'spontinifr'}, 'cz')
+    'spontini'
+
+    >>> get_name({'name':'spontini', 'name:en':'spontinien', 'name:fr':'spontinifr'}, 'fr')
+    'spontinifr'
+    """
+    name = properties.get(f"name:{lang}")
+    if name is None:
+        name = properties.get("name")
+    return name
